rhode: overlay: Align autobrightness calibration data with stock
Stock's framework overlays contain an array named
"config_screenExponentBacklight", which seems to be used as reference
data for autobrightness, instead of the AOSP standard
config_autoBrightnessDisplayValuesNits array. While the values in the
latter array are expressed in nits, the values in the former are simple
digital modulation values in the range 0-255. To map them into their
nits equivalents, a function was built from the pairs obtained by
coupling config_screenBrightnessBacklight with
config_screenBrightnessNits, using linear interpolation.

By using the obtained values, autobrightness seems to behave like on
stock.

Change-Id: Ic2b3ee7ac9c4857d0d58656385e012a3a41befd4